requirements.txt
datasets
google-genai
llm-api
re
